App test 리팩토링
비동기함수와 관련한 테스트를 actions test에서 하도록 변경하였습니다.
